+++ samples/src/main/java/com/example/TokenGenerationExample.java
@@ -10,7 +10,7 @@ public class TokenGenerationExample {
         // Generate BearerToken by specifying credentials.json file path
         try {
             String filePath = "<your_credentials.json_file_path>";
-            ResponseToken res = Token.GenerateBearerToken(filePath);
+            ResponseToken res = Token.generateBearerToken(filePath);
             System.out.println(res.getTokenType() + ":" + res.getAccessToken());
         } catch (SkyflowException e) {
             e.printStackTrace();
@@ -19,7 +19,7 @@ public class TokenGenerationExample {
         // Generate BearerToken by specifying credentials.json as string
         try {
             String filePath = "<your_credentials.json_file_as string>";
-            ResponseToken res = Token.GenerateBearerTokenFromCreds(filePath);
+            ResponseToken res = Token.generateBearerTokenFromCreds(filePath);
             System.out.println(res.getTokenType() + ":" + res.getAccessToken());
         } catch (SkyflowException e) {
             e.printStackTrace();